requirements.txt
openai~=0.27.4
fire~=0.5.0
GitPython~=3.1.31
requests~=2.28.0
